\name{highd2means-package}
\alias{highd2means-package}
\docType{package}
\title{
High-Dimensional Tests for two Population Mean Vectors
}
\description{
The package contains a few tests for the case of two high-dimensional population mean vectors. The user has the option to compute the asymptotic, the permutation or the bootstrap based p-value of the test.The tests are based on three other packages, namely the \pkg{highmean}, \pkg{highDmean} and the \pkg{PEtests}. We took the codes, modified them and made them more efficient.
}
\details{
\tabular{lll}{
Package: \tab highd2means\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-08-02\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Manos Papadakis \email{papadakm95@gmail.com}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Bai Z.D. and Saranadasa H. (1996). Effect of high dimension: by an example of a two sample
problem. Statistica Sinica, 6(2): 311--329.

Cai T.T., Liu W., and Xia Y. (2014). Two-sample test of high dimensional means under dependence.
Journal of the Royal Statistical Society: Series B (Statistical Methodology), 76(2): 349--372.

Chen S.X. and Qin Y.L. (2010). A two-sample test for high-dimensional data with applications to
gene-set testing. The Annals of Statistics, 38(2) 808--835.

Chen S.X., Li J., and Zhong P.S. (2014). Two-Sample Tests for High Dimensional Means with
Thresholding and Data Transformation. arXiv preprint arXiv:1410.2848.

Srivastava M.S. and Du M. (2008). A test for the mean vector with fewer observations than the
dimension. Journal of Multivariate Analysis, 99(3): 386--402.

Srivastava, M.S., Katayama, S., and Kano, Y. (2013). A two sample test in high dimensional data.
Journal of Multivariate Analysis, 114: 349--358.

Yu X., Li D., Xue L. and Li, R. (2023). Power-enhanced simultaneous test of high-dimensional
mean vectors and covariance matrices with application to gene-set testing. Journal of the American
Statistical Association, 118(544): 2548--2561.
}



